+++ tools/src/librtcltools/RtclStats.hpp
@@ -1,6 +1,6 @@
-// $Id: RtclStats.hpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RtclStats.hpp 1160 2019-06-07 17:30:17Z mueller $
 //
-// Copyright 2011-2013 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2011-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-07  1160   1.1    Rename Collect->Exec, not longer const
 // 2013-03-06   495   1.0.1  Rename Exec->Collect
 // 2011-02-26   364   1.0    Initial version
 // 2011-02-20   363   0.1    fFirst draft
@@ -49,8 +50,7 @@ namespace Retro {
       };
     
       static bool     GetArgs(RtclArgs& args, Context& cntx);
-      static bool     Collect(RtclArgs& args, const Context& cntx, 
-                              const Rstats& stats);
+      static bool     Exec(RtclArgs& args, const Context& cntx, Rstats& stats);
   };
 
 } // end namespace Retro