+++ tools/src/librw11/Rw11CntlRK11.hpp
@@ -1,6 +1,6 @@
-// $Id: Rw11CntlRK11.hpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11CntlRK11.hpp 1160 2019-06-07 17:30:17Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-07  1160   2.0.2  RdmaStats() not longer const
 // 2017-04-02   865   2.0.1  Dump(): add detail arg
 // 2015-01-03   627   2.0    use Rw11RdmaDisk
 // 2014-12-29   623   1.1    adopt to Rlink V4 attn logic
@@ -54,7 +55,7 @@ namespace Retro {
       void          SetChunkSize(size_t chunk);
       size_t        ChunkSize() const;
 
-      const Rstats& RdmaStats() const;
+      Rstats&       RdmaStats();
 
       virtual void  Dump(std::ostream& os, int ind=0, const char* text=0,
                          int detail=0) const;