+++ tools/src/librw11/Rw11Rdma.hpp
@@ -1,6 +1,6 @@
-// $Id: Rw11Rdma.hpp 1084 2018-12-16 12:23:53Z mueller $
+// $Id: Rw11Rdma.hpp 1160 2019-06-07 17:30:17Z mueller $
 //
-// Copyright 2015-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2015-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-07  1160   1.1.5  Stats() not longer const
 // 2018-12-16  1084   1.1.4  use =delete for noncopyable instead of boost
 // 2018-12-15  1083   1.1.3  for std::function setups: use rval ref and move
 // 2018-12-14  1081   1.1.2  use std::function instead of boost
@@ -72,7 +73,7 @@ namespace Retro {
       void          QueueWMem(uint32_t addr, const uint16_t* block, size_t size,
                               uint16_t mode);
 
-      const Rstats& Stats() const;
+      Rstats&       Stats();
       virtual void  Dump(std::ostream& os, int ind=0, const char* text=0,
                          int detail=0) const;