+++ tools/src/librwxxtpp/RtclRw11Cntl.cpp
@@ -1,4 +1,4 @@
-// $Id: RtclRw11Cntl.cpp 1114 2019-02-23 18:01:55Z mueller $
+// $Id: RtclRw11Cntl.cpp 1160 2019-06-07 17:30:17Z mueller $
 //
 // Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2019-06-07  1160   1.1.6  use RtclStats::Exec()
 // 2019-02-23  1114   1.1.5  use std::bind instead of lambda
 // 2018-12-17  1085   1.1.4  use std::lock_guard instead of boost
 // 2018-12-15  1082   1.1.3  use lambda instead of boost::bind
@@ -124,7 +125,7 @@ int RtclRw11Cntl::M_stats(RtclArgs& args)
 {
   RtclStats::Context cntx;
   if (!RtclStats::GetArgs(args, cntx)) return kERR;
-  if (!RtclStats::Collect(args, cntx, Obj().Stats())) return kERR;
+  if (!RtclStats::Exec(args, cntx, Obj().Stats())) return kERR;
   return kOK;
 }