+++ tools/src/librwxxtpp/RtclRw11CntlRL11.cpp
@@ -1,6 +1,6 @@
-// $Id: RtclRw11CntlRL11.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RtclRw11CntlRL11.cpp 1160 2019-06-07 17:30:17Z mueller $
 //
-// Copyright 2014-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2014-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,7 +13,8 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
-// 2017-04-16   878   1.2    add class in ctor; derive from RtclRw11CntlDiskBase
+// 2019-06-07  1160   1.1.1  use RtclStats::Exec()
+// 2017-04-16   878   1.1    add class in ctor; derive from RtclRw11CntlDiskBase
 // 2015-01-10   632   1.0    Initial version
 // 2014-06-08   561   0.1    First draft
 // ---------------------------------------------------------------------------
@@ -98,8 +99,8 @@ int RtclRw11CntlRL11::M_stats(RtclArgs& args)
 {
   RtclStats::Context cntx;
   if (!RtclStats::GetArgs(args, cntx)) return kERR;
-  if (!RtclStats::Collect(args, cntx, Obj().Stats())) return kERR;
-  if (!RtclStats::Collect(args, cntx, Obj().RdmaStats())) return kERR;
+  if (!RtclStats::Exec(args, cntx, Obj().Stats())) return kERR;
+  if (!RtclStats::Exec(args, cntx, Obj().RdmaStats())) return kERR;
   return kOK;
 }